[Security Solution] [Endpoint] Add new policy tabs layout (#110966)
* Add new policy tabs layout with policy settings and trusted apps tab. Only visible with feature flag enabled

* Add URL state of which tab is selected. Also refactored policy_details component. Hide sticky bar at bottom when trustedApps tab is selected

* Fix wrong constant path in routing

* Don't refresh policyItem if is not necessary

* Remove old code and use new form layout component even if FF is disabled

* Split test file

* Clean test file with unused mocks

* Fixes failing test

* Address pr comments
