+++ x-pack/plugins/security_solution/public/management/pages/policy/store/policy_details/middleware.ts
@@ -13,6 +13,7 @@ import {
   isOnPolicyDetailsPage,
   policyDetails,
   policyDetailsForUpdate,
+  needsToRefresh,
 } from './selectors';
 import {
   sendGetPackagePolicy,
@@ -31,7 +32,7 @@ export const policyDetailsMiddlewareFactory: ImmutableMiddlewareFactory<PolicyDe
     next(action);
     const state = getState();
 
-    if (action.type === 'userChangedUrl' && isOnPolicyDetailsPage(state)) {
+    if (action.type === 'userChangedUrl' && needsToRefresh(state) && isOnPolicyDetailsPage(state)) {
       const id = policyIdFromParams(state);
       let policyItem: PolicyData;
