+++ x-pack/plugins/security_solution/public/management/pages/policy/store/policy_details/selectors.ts
@@ -18,7 +18,10 @@ import {
   UIPolicyConfig,
 } from '../../../../../../common/endpoint/types';
 import { policyFactory as policyConfigFactory } from '../../../../../../common/endpoint/models/policy_config';
-import { MANAGEMENT_ROUTING_POLICY_DETAILS_PATH } from '../../../../common/constants';
+import {
+  MANAGEMENT_ROUTING_POLICY_DETAILS_FORM_PATH,
+  MANAGEMENT_ROUTING_POLICY_DETAILS_TRUSTED_APPS_PATH,
+} from '../../../../common/constants';
 import { ManagementRoutePolicyDetailsParams } from '../../../../types';
 import { getPolicyDataForUpdate } from '../../../../../../common/endpoint/service/policy/get_policy_data_for_update';
 
@@ -70,15 +73,36 @@ export const policyDetailsForUpdate: (
   }
 });
 
-/** Returns a boolean of whether the user is on the policy details page or not */
-export const isOnPolicyDetailsPage = (state: Immutable<PolicyDetailsState>) => {
+/**
+ * Checks if data needs to be refreshed
+ */
+export const needsToRefresh = (state: Immutable<PolicyDetailsState>): boolean => {
+  return !state.policyItem && !state.apiError;
+};
+
+/** Returns a boolean of whether the user is on the policy form page or not */
+export const isOnPolicyFormPage = (state: Immutable<PolicyDetailsState>) => {
   return (
     matchPath(state.location?.pathname ?? '', {
-      path: MANAGEMENT_ROUTING_POLICY_DETAILS_PATH,
+      path: MANAGEMENT_ROUTING_POLICY_DETAILS_FORM_PATH,
       exact: true,
     }) !== null
   );
 };
+
+/** Returns a boolean of whether the user is on the policy details page or not */
+export const isOnPolicyTrustedAppsPage = (state: Immutable<PolicyDetailsState>) => {
+  return (
+    matchPath(state.location?.pathname ?? '', {
+      path: MANAGEMENT_ROUTING_POLICY_DETAILS_TRUSTED_APPS_PATH,
+      exact: true,
+    }) !== null
+  );
+};
+
+/** Returns a boolean of whether the user is on some of the policy details page or not */
+export const isOnPolicyDetailsPage = (state: Immutable<PolicyDetailsState>) =>
+  isOnPolicyFormPage(state) || isOnPolicyTrustedAppsPage(state);
 
 /** Returns the license info fetched from the license service */
 export const license = (state: Immutable<PolicyDetailsState>) => {
@@ -91,7 +115,10 @@ export const policyIdFromParams: (state: Immutable<PolicyDetailsState>) => strin
   (location: PolicyDetailsState['location']) => {
     return (
       matchPath<ManagementRoutePolicyDetailsParams>(location?.pathname ?? '', {
-        path: MANAGEMENT_ROUTING_POLICY_DETAILS_PATH,
+        path: [
+          MANAGEMENT_ROUTING_POLICY_DETAILS_FORM_PATH,
+          MANAGEMENT_ROUTING_POLICY_DETAILS_TRUSTED_APPS_PATH,
+        ],
         exact: true,
       })?.params?.policyId ?? ''
     );
