+++ x-pack/plugins/security_solution/public/management/pages/policy/view/trusted_apps/layout/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { PolicyTrustedAppsLayout } from './policy_trusted_apps_layout';
